Finished writing HashTable class
Hash class is a basic hash table using djb2 for the hash function and chaining to handle collisions.
despite using a fairly efficient algorithm for hashing, it is still recommended to use > 2 times
as many spots as there are data members. This shouldn't be a problem memory wise, as space is
only allocated as it is needed (an array of pointers is allocated statically, however, so the minimum space required will be something like 8*table_size bytes)
The table handles reading and writing to a binary file, table.bin, on its own,
so if the file is in the directory of the program using that hash table, it will read from the file
and construct the new table with its contents. When the table's destructor is called, it will
first write the topic data to the binary file, then delete the heap space, and the compiler
handles deleting the array of pointers and integer fields.